% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{calc_coverage}
\alias{calc_coverage}
\title{Function that computes how often the true value of the parameter is included within the 
90/10 high posterior density interval}
\usage{
calc_coverage(obj, rep = 1, quantiles = c(0.95, 0.05))
}
\arguments{
\item{est_param}{A matrix of posterior draws of a parameter}

\item{true_param}{A matrix (one column) of true parameter values}
}
\description{
Function that computes how often the true value of the parameter is included within the 
90/10 high posterior density interval
}

