% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{calc_resid}
\alias{calc_resid}
\title{Residual function for checking estimated samples compared to true simulation scores
Returns a data frame with residuals plus quantiles.}
\usage{
calc_resid(obj, rep = 1)
}
\arguments{
\item{est_param}{A matrix of posterior draws of a parameter}

\item{true_param}{A matrix (one column) of true parameter values}
}
\description{
Residual function for checking estimated samples compared to true simulation scores
Returns a data frame with residuals plus quantiles.
}
