% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{compare_models}
\alias{compare_models}
\title{Function to compare two fitted idealstan models by plotting ideal points. Assumes that underlying data
is the same for both models.
 @export}
\usage{
compare_models(model1 = NULL, model2 = NULL)
}
\description{
Function to compare two fitted idealstan models by plotting ideal points. Assumes that underlying data
is the same for both models.
 @export
}

