% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.calc_true_pts}
\alias{.calc_true_pts}
\title{used to calculate the true ideal points
given that a non-centered parameterization is used.}
\usage{
.calc_true_pts(obj)
}
\description{
used to calculate the true ideal points
given that a non-centered parameterization is used.
}
