% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.check_quoted}
\alias{.check_quoted}
\title{Simple function to test for what an input is
Default_val should be quoted}
\usage{
.check_quoted(quoted = NULL, default_val)
}
\description{
Simple function to test for what an input is
Default_val should be quoted
}
