% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.constrain_fix}
\alias{.constrain_fix}
\title{Function that constrains certain known parameters}
\usage{
.constrain_fix(object = NULL, restrict_ind_high = NULL,
  restrict_ind_low = NULL, use_groups = NULL, ...)
}
\description{
Function that constrains certain known parameters
}
