% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.create_array}
\alias{.create_array}
\title{Helper function to create arrays}
\usage{
.create_array(input_matrix, arr_dim = 2, row_var = NULL,
  col_var_name = NULL, col_var_value, third_dim_var = NULL)
}
\arguments{
\item{input_matrix}{Either a data.frame in long mode or a single matrix}

\item{arr_dim}{If \code{input_matrix} is a single matrix, \code{arr_dim} determines the length of the resulting array}

\item{row_var}{Unquoted variable name that identifies the data.frame column corresponding to the rows (1st dimension) of the array (must be unique)}

\item{col_var_name}{Unquoted variable name that identifies the data.frame column corresponding names of the columns (2nd dimension) of the array}

\item{col_var_value}{Unquoted variable name that identifies the data.frame column corresponding to the values that populate the cells of the array}

\item{third_dim_var}{Unquoted variable name that identifis the data.frame column corresponding to the dimension around which to stack the matrices (3rd dimension of array)}
}
\description{
Function takes a data.frame in long mode and converts it to an array. Function can also repeat a 
single matrix to fill out an array.
}
