% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.det_missing}
\alias{.det_missing}
\title{Helper function that determines if there are more missing than non-missing
observations}
\usage{
.det_missing(object, model_type = NULL)
}
\description{
Helper function that determines if there are more missing than non-missing
observations
}
