% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.extract_nonp}
\alias{.extract_nonp}
\title{a slightly hacked function to extract parameters as I want to}
\usage{
.extract_nonp(object, pars, permuted = TRUE, inc_warmup = FALSE,
  include = TRUE)
}
\description{
a slightly hacked function to extract parameters as I want to
}
