% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP_func.R
\name{.gen_ts_data}
\alias{.gen_ts_data}
\title{Function to generate random-walk or AR(1) person parameters
Recursively generate data}
\usage{
.gen_ts_data(t, adj_in, alpha_int, sigma, init_sides)
}
\description{
Function to generate random-walk or AR(1) person parameters
Recursively generate data
}
