% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.get_kept_samples2}
\alias{.get_kept_samples2}
\title{we are going to modify this rstan function so that it no longer permutes
just delete the last term -- maybe submit PR to rstan}
\usage{
.get_kept_samples2(n, sim)
}
\description{
we are going to modify this rstan function so that it no longer permutes
just delete the last term -- maybe submit PR to rstan
}
