% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.init_stan}
\alias{.init_stan}
\title{Simple function to provide initial values to Stan given current values of restrict_sd}
\usage{
.init_stan(chain_id = NULL, restrict_sd = NULL, person_sd = NULL,
  num_legis = NULL, diff_high = NULL, actual = TRUE)
}
\description{
Simple function to provide initial values to Stan given current values of restrict_sd
}
