% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.item_plot_binary}
\alias{.item_plot_binary}
\title{Generate item-level midpoints for binary IRT outcomes}
\usage{
.item_plot_binary(param_name, object, high_limit = NULL,
  low_limit = NULL, all = FALSE, aggregate = FALSE)
}
\description{
Generate item-level midpoints for binary IRT outcomes
}
