% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.item_plot_ls}
\alias{.item_plot_ls}
\title{Generate item-level midpoints for binary latent-space outcomes}
\usage{
.item_plot_ls(param_name, object, high_limit = NULL, low_limit = NULL)
}
\description{
Generate item-level midpoints for binary latent-space outcomes
}
