% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.pinned_fix}
\alias{.pinned_fix}
\title{Function that pins certain parameters to fixed points}
\usage{
.pinned_fix(object = NULL, nfix = NULL, restrict_params = NULL,
  restrict_ind_high = NULL, use_groups = NULL, ...)
}
\description{
Function that pins certain parameters to fixed points
}
