% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predict.R
\name{.predict_2pl}
\alias{.predict_2pl}
\title{Function to predict ordinary 2 PL models}
\usage{
.predict_2pl(all_params = NULL, n_iters = NULL, sample_draws = NULL,
  sample_scores = NULL, legis_points = NULL, bill_points = NULL,
  time_points = NULL, obj = NULL)
}
\description{
Function to predict ordinary 2 PL models
}
