% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predict.R
\name{.predict_abs_bin}
\alias{.predict_abs_bin}
\title{Function to predict absence-inflated binary models}
\usage{
.predict_abs_bin(all_params = NULL, n_iters = NULL,
  sample_draws = NULL, sample_scores = NULL, legis_points = NULL,
  bill_points = NULL, time_points = NULL, y = NULL, obj = NULL)
}
\description{
Function to predict absence-inflated binary models
}
