% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predict.R
\name{.predict_abs_ord_ll}
\alias{.predict_abs_ord_ll}
\title{Function to calculate log-likelihood of absence-inflated ordinal models}
\usage{
.predict_abs_ord_ll(all_params = NULL, n_iters = NULL,
  sample_draws = NULL, sample_scores = NULL, legis_points = NULL,
  bill_points = NULL, time_points = NULL, y = NULL, obj = NULL)
}
\description{
Function to calculate log-likelihood of absence-inflated ordinal models
}
