% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.prepare_legis_data}
\alias{.prepare_legis_data}
\title{Helper function for preparing person ideal point plot data}
\usage{
.prepare_legis_data(object)
}
\description{
Helper function for preparing person ideal point plot data
}
