% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.prior_fit}
\alias{.prior_fit}
\title{Function that uses a previously-identified model to maintain comparability across model types}
\usage{
.prior_fit(object = NULL, prior_fit = NULL, use_groups = NULL, ...)
}
\description{
Function that uses a previously-identified model to maintain comparability across model types
}
