% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predict.R
\name{.sample_cut}
\alias{.sample_cut}
\title{Helper function for sampling from ordinal cutpoints}
\usage{
.sample_cut(pr_vote = NULL, cutpoints = NULL, n_outcomes = NULL)
}
\description{
Helper function for sampling from ordinal cutpoints
}
