% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_predict.R
\name{.sample_cut_ll}
\alias{.sample_cut_ll}
\title{Helper function for obtaining density given ordinal cutpoints}
\usage{
.sample_cut_ll(pr_vote = NULL, cutpoints = NULL, n_outcomes = NULL)
}
\description{
Helper function for obtaining density given ordinal cutpoints
}
