% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{extract_log_lik.idealstan}
\alias{extract_log_lik.idealstan}
\title{This function returns a matrix of an S by N matrix of draws from a fitted idealstan model, where S is the size of the 
posterior sample and N is the total number of parameters in the idealstan model. This matrix can then be used to 
fit an information criterion to assess model fit, see the loo package for details.}
\usage{
\method{extract_log_lik}{idealstan}(object)
}
\description{
This function returns a matrix of an S by N matrix of draws from a fitted idealstan model, where S is the size of the 
posterior sample and N is the total number of parameters in the idealstan model. This matrix can then be used to 
fit an information criterion to assess model fit, see the loo package for details.
}
