% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{extract_log_lik}
\alias{extract_log_lik}
\title{Generic Method for Extracting Log Likelihood from Stan Objects}
\usage{
extract_log_lik(object, ...)
}
\description{
This function is a generic that is used to match the functions used with \code{\link[loo]{loo}} to calculate
Bayesian information criteria on models.
}
