% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{id_estimate}
\alias{id_estimate}
\title{Estimate an \code{idealstan} model}
\usage{
id_estimate(idealdata = NULL, model_type = 2, use_subset = FALSE,
  sample_it = FALSE, subset_party = NULL, subset_legis = NULL,
  sample_size = 20, nchains = 4, niters = 2000, use_vb = FALSE,
  nfix = 1, restrict_params = "legis", pin_vals = NULL,
  restrict_ind_high = NULL, restrict_ind_low = NULL,
  restrict_type = "constrain_oneway", fixtype = "vb",
  warmup = floor(niters/2), ncores = 4, auto_id = FALSE,
  discrim_reg_sd = 2, discrim_abs_sd = 2, legis_sd = 1, diff_reg_sd = 5,
  diff_abs_sd = 5, restrict_sd = 1, ...)
}
\arguments{
\item{model_type}{An integer reflecting the kind of model to be estimated. See above.}

\item{use_subset}{Whether a subset of the legislators/persons should be used instead of the full response matrix}

\item{sample_it}{Whether or not to use a random subsample of the response matrix. Useful for testing.}

\item{subset_party}{If legislative/person data was included in the \code{\link{id_make}} function, then you can subset by
any value in the \code{$bloc} column of that data if \code{use_subset} is \code{TRUE}.}

\item{subset_legis}{A list of character values of names of legislators/persons to use to subset if \code{use_subset} is 
\code{TRUE} and legislative/person data was included in the \code{\link{id_make}} function with the required \code{$legis.names}
column}

\item{sample_size}{If \code{sample_it} is \code{TRUE}, this value reflects how many legislators/persons will be sampled from
the response matrix}

\item{nchains}{The number of chains to use in Stan's sampler. Minimum is one. See \code{\link[rstan]{stan}} for more info.}

\item{niters}{The number of iterations to run Stan's sampler. Shouldn't be set much lower than 500. See \code{\link[rstan]{stan}} for more info.}

\item{use_vb}{Whether or not to use Stan's variational Bayesian inference engine instead of full Bayesian inference. Pros: it's much faster.
Cons: it's not quite as accurate. See \code{\link[rstan]{vb}} for more info.}

\item{restrict_params}{Determines which parameters in the model are used for identification. If \code{fixtype} is 'vb', can be
either 'legis' (legislators/persons) or 'bills' (discrimination parameters for bills/items). Otherwise, the value must be one
of the available classes of parameters: 'legis' for persons/legislators, 'discrim_abs' for the absence-inflated bill discrimination
parameters, or 'discrim_reg' for the non-inflated bill discrimination parameters.}

\item{pin_vals}{If \code{fixtype} is set to 'pinned', then a vector of numeric values of which to pin the legislators/persons or 
bills/items should be given.}

\item{restrict_ind_high}{If \code{fixtype} is not "vb", the particular indices of legislators/persons or bills/items to constrain high}

\item{restrict_ind_low}{If \code{fixtype} is not "vb", the particular indices of legislators/persons or bills/items to constrain low. 
(Note: not used if values are pinned).}

\item{restrict_type}{Determines whether constraints will only be set on one direction 'constrain_oneway' or both high/low
constrains 'constrain_twoway'. Ignored if \code{fixtype} is equal to 'pinned'.}

\item{fixtype}{Sets the particular kind of identification used on the model, could be one of 'vb','constrained' or 'pinned'.
See details for more information.}

\item{warmup}{The number of iterations to use to calibrate Stan's sampler on a given model. Shouldn't be less than 100. 
See \code{\link[rstan]{stan}} for more info.}

\item{ncores}{The number of cores in your computer to use for parallel processing in the Stan engine. 
See \code{\link[rstan]{stan}} for more info.}

\item{auto_id}{If \code{TRUE}, a variational Bayesian version of your model will be run using \code{\link[rstan]{vb}} in a 
non-identified form. Depending on what is set via \code{restrict_type} and \code{restrict_params}, bills/items or 
persons/legislators will be used to constrain and identify the model.}

\item{discrim_reg_sd}{Set the prior standard deviation of the bimodal prior for the discrimination parameters for the non-inflated model.}

\item{discrim_abs_sd}{Set the prior standard deviation of the bimodal prior for the discrimination parameters for the inflated model.}

\item{legis_sd}{Set the prior standard deviation for the legislators (persons) parameters}

\item{diff_reg_sd}{Set the prior standard deviation for the bill (item) intercepts for the non-inflated model.}

\item{diff_abs_sd}{Set the prior standard deviation for the bill (item) intercepts for the inflated model.}

\item{restrict_sd}{Set the prior standard deviation for constrained parameters}

\item{...}{Additional parameters passed on to Stan's sampling engine. See \code{\link[rstan]{stan}} for more information.}

\item{auto_id}{If \code{fixtype} is set to "vb", the identification function will be re-run 
with multiple \code{\link[rstan]{vb}} models until the different model estimates show underlying identification. Not perfect, but
works pretty well.}
}
\value{
A fitted \code{\link{idealstan}} object that contains posterior samples of all parameters either via full Bayesian infererence
or a variational approximation if \code{use_vb} is set to \code{TRUE}. This object can then be passed to the plotting functions for further analysis.
}
\description{
This function will take a pre-processed \code{\link[=id_make]{idealdata}} vote/score matrix and run one of the available IRT models on the data using
Stan's MCMC engine.
}
\details{
To run an IRT ideal point model, you must first pre-process your data using the \code{\link{id_make}} function. Be sure to specify the correct options for the
kind of model you are going to run: if you want to run an ordinal and/or an inflated model, the data needs to be processed differently.
As of this version of \code{idealstan}, the following model types are available:
\enumerate{
  \item IRT 2-PL (binary response) ideal point model, no absence inflation
  \item IRT 2-PL model (binary response) with absence inflation
  \item Ordinal IRT (rating scale) model incorporating abstentions as middle category, no absence inflation
  \item Ordinal IRT (rating scale) model incorporating abstentions as middle category with absence inflation
}
Additional models are available but have not yet been tested. You can find them by searching the included stan files for more info.
}
\section{Identification}{

Identifying IRT models is challenging, and ideal point models are still more challenging because the discrimination parameters are not constrained.
As a result, more care must be taken to obtain estimates that are the same regardless of starting values. 
The parameter \code{fixtype} enables you to change the type of identification used. The default, 'vb', does not require any further
information from you in order for the model to be fit. In this version of identification, an unidentified model is run using
variational Bayesian inference (see \code{\link[rstan]{vb}}). If \code{restrict_type} is set to 'constrain_oneway', then the \code{nfix}
highest legislators/persons (if \code{restrict_params} is 'legis') or bills/items (if \code{restrict_params} equals 'bills') 
are used to constrain and identify the model. If \code{restrict_type} is set to 'constrain_twoway', then \code{nfix} highest and lowest legislators/persons
or bills/items are used to constrain and identify the model.
In addition, if \code{fixtype='vb'} is used, \code{auto_id} can be set to \code{TRUE}. This will run additional varitional Bayesian models
using the identification achieved and will see if the signs of the estimated parameters are at least 90 percent in the same direction. If so, 
the model is considered identified. If not, the function will re-run and will increase \code{nfix} by one to see if that will identify the model, 
ad infinitum.
If \code{fixtype} is set to 'constrained', then identification is achieved by constraining a specified list of legislators/persons or
bills/items. If \code{restrict_type} is 'constrain_oneway', then the indices of all constrained legislators/persons or bills/items
should be included as the row or column indices of these parameters in the response matrix as a vector in \code{restrict_ind_high}.
If \code{restrict_type} is 'constrain_twoway,' then the indices of high constrained parameters should go in 
\code{restrict_ind_high} and the indices of low constrained parameters in \code{restrict_ind_low}. The numbers of high and low
constrained parameters should be equal. To pick the parameter to constrain, set \code{restrict_params} to 'legis' for legislators/persons,
'discrim_abs' for absence-inflated bill discrimination parameters, and 'discrim_reg' for non-inflated bill discrimination parameters.
If \code{fixtype} is set to 'pinned', then identification is achieved via pinning with very tight priors a set of legislators/persons or bills/items.
The indices of legislators/persons (i.e. row indices in the response matrix) or bills/items (column indices in the response matrix) should be passed as a vector
to \code{restrict_ind_high}, while \code{restrict_ind_low} should be left blank. The particular values to pin these parameters is passed as a 
numeric vector to \code{pin_vals}.
}

\seealso{
\code{\link{id_make}} for pre-processing data,
\code{\link{id_plot}} for plotting results,
\code{\link{summary}} for obtaining posterior quantiles,
\code{\link{posterior_predict}} for producing predictive replications.
}
