% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{id_load_mpi}
\alias{id_load_mpi}
\title{Load Results from CMDstan to Idealstan Object}
\usage{
id_load_mpi(folder = NULL, samples = NULL)
}
\arguments{
\item{folder}{The location of the folder (full file path) 
that was given to the \code{id_estimate} function to 
export idealstan data}

\item{samples}{The full file path to the CSV output
from the successful CMDstan run}
}
\description{
Given the location of the folder of seralized data from id_estimate,
and the name of the CSV output from a CMDstan run, 
this function will load the original idealstan data and the 
Markov samples to create an idealstan object that can be 
used for further analysis with idealstan functions.
}
