% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{id_make}
\alias{id_make}
\title{Create data to run IRT model}
\usage{
id_make(vote_data = NULL, simul_data = NULL, legis_cov = NULL,
  bill_cov_reg = NULL, bill_cov_abs = NULL, legis_data = NULL,
  bill_data = NULL, abs_vote = NA, yes_vote = 3L, no_vote = 1L,
  abst_vote = 2L, ordinal = TRUE, time = NULL, exclude_level = NULL,
  inflate = TRUE, simulation = FALSE)
}
\arguments{
\item{vote_data}{Your legislator-bill (person-item) matrix in which legislators (persons) are in rows and 
  bills (items) are in columns. The cells of the matrix should contain consecutive integers \eqn{1,...K} for 
  ordinal \eqn{K} outcomes or integers \eqn{[0,1]} for a binary outcome in which 0 equals no (or incorrect) and 
  1 equals yes (or correct). If absences are included, they should be coded as the highest category \eqn{K+1} for
  ordinal outcomes and 2 for binary outcomes.
@param simul_data Optionally, data that has been generated by the \code{\link{id_sim}} function.
@param legis_cov A matrix of legislator covariates for hierarchical modeling. If multiple time points are used, then should be a 
multi-column matrix with one column for each time point.
@param bill_cov_reg A one-column matrix of bill covariates for hierarchical modeling of bill discrimination parameters in the non-inflated model
@param bill_cov_abs A one-column matrix of bill covariates for hierarchical modeling of bill discrimination parameters in the inflated model
@param legis_data An optional data frame of information about the legislators (persons). Should include a \code{legis.names} column containing names of
legislators (persons) and optionally a \code{bloc} column with names of parties/blocs in the legislature
@param bill_data An optional data frame of bill/item labels and other information. Should include \code{bill.names} column 
containing bill labels; other columns are optioanl.
@param abs_vote The value (numeric or character) that indicate absences in the data. If there are multiple possible values, 
pass along a numeric or character vector of all such values.
@param yes_vote The value (numeric or character) that indicate yes/correct responses in the data.
If there are multiple possible values, 
pass along a numeric or character vector of all such values.
@param no_vote The value (numeric or character) that indicate no/incorrect responses in the data.
If there are multiple possible values, 
pass along a numeric or character vector of all such values.
@param abst_vote The value (numeric or character) that indicate abstentions (a middle ordinal category) in the data. 
If there are multiple possible values, 
pass along a numeric or character vector of all such values.
@param ordinal Whether or not the data contain ordinal responses. If \code{TRUE}, abstentions are used as a middle category in constructing the outcome.
Otherwise the response is assumed to be binary (yes/no) or (correct/incorrect).
@param time An optional integer vector of length equal to the number of columns in the response matrix (i.e., the number of bills or items)
that indicates for each bill/item which time point that bill/item belongs to. This time vector is only used for models in which legislator/person
ideal points are allowed to vary over time. 
@param exclude_level A vector of any values that should be treated as \code{NA} in the response matrix
@param inflate If \code{TRUE}, the response matrix is set up to enable modeling of absences as an inflation model in \code{\link{id_estimate}}
@param simulation If \code{TRUE}, simulated values are saved in the \code{idealdata} object for later plotting with the \code{\link{id_plot_sim}} function}
}
\value{
A \code{\link{idealdata}} object that can then be used in the \code{\link{id_estimate}} function to fit a model.
}
\description{
To run an IRT model using \code{idealstan}, you must first process your data using the \code{id_make} function.
}
