% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{id_me,idealstan-method}
\alias{id_me,idealstan-method}
\title{Calculate ideal point marginal effects}
\usage{
\S4method{id_me}{idealstan}(
  object,
  covariate = NULL,
  group_effects = NULL,
  pred_outcome = NULL,
  eps = 1e-04,
  draws = 100,
  cores = 1,
  lb = 0.025,
  upb = 0.975,
  ...
)
}
\arguments{
\item{object}{A fitted \code{idealstan} model}

\item{covariate}{The character value for a covariate passed to the
\code{id_make} function before model fitting. Only one covariate can be processed
at a time.}

\item{group_effects}{character value of a covariate included in the formula passed
to \code{id_make} for which marginal effect summaries should be grouped by. Useful
when looking at the marginal effect of an interaction. Note that grouping by a covariate
with many values will result in slow performance.}

\item{pred_outcome}{Numeric value for level of outcome to predict for ordinal responses.
Defaults to top level.}

\item{eps}{Parameter for numerical differentiation (usually does not need to be changed)
passed on to \link{id_post_pred}}

\item{draws}{The total number of draws to use when calculating the marginal effects.
Defaults to 100. Use option "all" to use all available MCMC draws.}

\item{cores}{The total number of cores to use when calculating the marginal effects.
Defaults to 1.}

\item{lb}{The quantile for the lower bound of the aggregated effects (default is 0.025 for a 95\% interval)}

\item{upb}{The quantile for the upper bound of the aggregated effects (default is 0.975 for a 95\% interval)}

\item{...}{Additional arguments passed on to \link{id_post_pred}}
}
\value{
A list with two objects, \code{ideal_effects} with one estimate of the
marginal effect per item and posterior draw and \code{sum_ideal_effects} with
one row per item with that item's median ideal point marginal effect with the quantiles
defined by the \code{upb} and \code{lb} parameters.
}
\description{
This function allows you to calculate ideal point marginal effects for a
given person-level hierarchical covariate.
}
\details{
This function will calculate item-level ideal point marginal effects
for a given covariate that was passed to the \code{id_make} function using the
\code{person_cov} option. The function will iterate over all items in the model
and use numerical differentiation to calculate responses in the scale of the
outcome for each item. Note: if the covariate is binary (i.e., only has two values),
then the function will calculate the difference between these two values instead of
using numerical differentation.
}
