% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{id_params}
\alias{id_params}
\title{Function that works with id_model to re-arrange bills or legislators to fix/constrain for identification}
\usage{
id_params(lookat_params = NULL, fixparams = NULL, nfix = NULL, x = NULL)
}
\description{
Function that works with id_model to re-arrange bills or legislators to fix/constrain for identification
}

