% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{id_plot}
\alias{id_plot}
\title{Generic Function for Plotting \code{idealstan} objects}
\usage{
id_plot(object, ...)
}
\arguments{
\item{object}{An \code{idealstan} object}
}
\description{
This generic function will run all the plotting functions associated with fitted \code{idealstan} objects.
}
