% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{id_plot_gbeta_prior}
\alias{id_plot_gbeta_prior}
\title{Launch the Generalized Beta Distribution Explorer Shiny App}
\usage{
id_plot_gbeta_prior(initial_y = 0, prior_sample_size = 200, limits = c(-1, 1))
}
\arguments{
\item{initial_y}{Initial observed variate \code{y} (default: 0)}

\item{prior_sample_size}{Relative level of precision/tightness around \code{y} (default: 200). Can be thought of as the relative sample size used to estimate \code{y}.}

\item{limits}{Limits of prior. Default is \verb{[-1,1]}. Set to \verb{[0,1]} for a conventional IRT with all positive discrimination parameters (i.e. as in a test-taking scenario).}
}
\description{
This function starts an interactive Shiny application to visualize a generalized Beta distribution
on a symmetrical interval from -scale to +scale in order to calculate the distribution paramters alpha (the \code{restrict_sd_high}/\code{restrict_N_low} parameter to \code{id_estimate}) and beta (the \code{restrict_N_high}/\code{restrict_sd_low} parameter to \code{id_estimate}). This function is useful for understanding what values to use to pin item discrimination parameters in \code{id_estimate} given a specific prior mean \code{initial_y} and a relative level of precision \code{prior_sample_size} (also known as \code{phi}). Higher values of \code{prior_sample_size} will imply a tighter prior around the pinned discrimination parameter.
}
\details{
The function is also useful for calculating the prior for all non-constrained discrimination parameters as well (\code{discrim_reg_shape} and \code{discrim_reg_scale}). These parameters are denoted in the Shiny app output for easy cut and paste to the \code{id_estimate} function call.
}
