% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{id_plot_legis}
\alias{id_plot_legis}
\title{Plot Legislator/Person and Bill/Item Ideal Points}
\usage{
id_plot_legis(object, return_data = FALSE, bill_plot = NULL,
  text_size_label = 2, text_size_party = 2.5, hjust_length = -0.7,
  legis_labels = TRUE, legis_ci_alpha = 0.5, abs_and_reg = "both",
  show_true = FALSE, party_color = TRUE, party_overlap = FALSE, ...)
}
\arguments{
\item{return_data}{If true, the calculated legislator/bill data is returned along with the plot in a list}

\item{bill_plot}{The column index of the bill/item midpoint to overlay on the plot}

\item{text_size_label}{ggplot2 text size for legislator labels}

\item{text_size_party}{ggplot2 text size for party text used for points}

\item{hjust_length}{horizontal adjustement of the legislator labels}

\item{legis_labels}{if \code{TRUE}, use the legis.names column to plot legislator labels}

\item{legis_ci_alpha}{The transparency level of the dot plot and confidence bars}

\item{abs_and_reg}{Whether to show both absence and regular midpoints if the model is absence-inflated}

\item{show_true}{Whether to show the true values of the legislators (if model has been simulated)}

\item{party_color}{If \code{TRUE}, give each party/bloc a different color}

\item{party_overlap}{Whether to prevent the text from overlapping itself (ggplot2 option)}
}
\description{
This function can be used on a fitted \code{\link{idealstan}} object to plot the relative positions and 
uncertainties of legislator/persons and bills/items.
}
\details{
This plot shows the distribution of ideal points for the legislators/persons in the model. It will plot them as a vertical
dot plot with associated high-density posterior interval (10\% to 90\%). In addition, if the column index for a 
bill/item from the response matrix is passed to the \code{bill_plot} option, then a bill midpoint will be overlain
on the ideal point plot, showing the point at which legislators/persons are indifferent to voting/answering on the 
bill/item. Note that because this is an ideal point model, it is not possible to tell from the midpoint itself
which side will be voting which way. For that reason, the legislators/persons are colored by their votes/scores to
make it clear.
}
