% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{id_plot_legis_dyn}
\alias{id_plot_legis_dyn}
\title{Plot Dynamic Person Ideal Points (Deprecated)}
\usage{
id_plot_legis_dyn(
  object,
  return_data = FALSE,
  include = NULL,
  item_plot = NULL,
  text_size_label = 2,
  text_size_group = 2.5,
  high_limit = 0.95,
  low_limit = 0.05,
  line_size = 1,
  highlight = NULL,
  plot_text = TRUE,
  use_ci = TRUE,
  plot_lines = 0,
  draw_line_alpha = 0.5,
  person_line_alpha = 0.3,
  person_ci_alpha = 0.8,
  item_plot_type = "non-inflated",
  show_true = FALSE,
  group_color = TRUE,
  hpd_limit = 10,
  sample_persons = NULL,
  use_chain = NULL,
  add_cov = TRUE,
  ...
)
}
\arguments{
\item{object}{A fitted \code{idealstan} object or a named list of \code{idealstan}
objects if the plot is supposed to show a comparison of different fitted \code{idealstan}
models (see Time Series vignette)}

\item{return_data}{If true, the calculated legislator/bill data is returned along with the plot in a list}

\item{include}{Specify a list of person/legislator IDs to include in the plot (all others excluded)}

\item{item_plot}{The value of the item/bill for which to plot its midpoint (character value)}

\item{text_size_label}{ggplot2 text size for legislator labels}

\item{text_size_group}{ggplot2 text size for group text used for points}

\item{high_limit}{A number between 0 and 1 showing the upper limit to compute the
posterior uncertainty interval (defaults to 0.95).}

\item{low_limit}{A number between 0 and 1 showing the lower limit to compute the
posterior uncertainty interval (defaults to 0.05).}

\item{line_size}{Sets the size of the line of the time-varying ideal points.}

\item{highlight}{A character referring to one of the persons in \code{person_labels} that the plot can highlight relative to other persons}

\item{plot_text}{If \code{TRUE}, will plot \code{person_labels} over the lines.}

\item{use_ci}{Whether or not high-posterior density intervals (credible intervals) should be
plotted over the estimates (turn off if the plot is too busy)}

\item{plot_lines}{The number of lines of actual draws of time-varying ideal points
to draw on the plot. Note that these are grouped by persons. Specific draws selected at random
from total number of draws of the estimation. Default is 0.}

\item{draw_line_alpha}{The opacity of lines plotted over the distribution (should be
between 0 and 1, default is 0.5).}

\item{person_line_alpha}{The transparency level of the time-varying ideal point line}

\item{person_ci_alpha}{The transparency level of ribbon confidence interval around the time-varying ideal points}

\item{item_plot_type}{Whether to show the \code{'non-inflated'} item/bill midpoints,
the \code{'inflated'} item/bill midpoints, or produce plots for \code{'both'} kinds of models.
Defaults to \code{'non-inflated'} and will only display an item/bill midpoint if one has been
specified in \code{item_plot}.}

\item{show_true}{Whether to show the true values of the legislators (if model has been simulated)}

\item{group_color}{If \code{TRUE}, use the groups instead of individuals to plot colours}

\item{hpd_limit}{The greatest absolute difference in high-posterior density interval shown for any point. Useful for excluding imprecisely estimated persons/legislators from the plot. Leave NULL if you don't want to exclude any.}

\item{sample_persons}{If you don't want to use the full number of persons/legislators from the model, enter a proportion (between 0 and 1) to select
only a fraction of the persons/legislators.
(see \code{\link[=id_sim_gen]{id_sim_gen()}})}

\item{use_chain}{ID of MCMC chain to use rather than combining all chains.
Default is NULL which will use all chains and is recommended.}

\item{add_cov}{Whether to add values of hierarchical person-level covariates to the
time trends (defaults to TRUE).}

\item{...}{Other options passed on to plotting function, currently ignored}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{id_plot_legis_dyn()} was renamed to \code{\link[=id_plot_persons_dyn]{id_plot_persons_dyn()}} for clarity and consistency.
}
\keyword{internal}
