% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{id_plot_sims}
\alias{id_plot_sims}
\alias{id_show_trues}
\title{This function plots the results from a simulation generated by \code{\link[=id_sim_gen]{id_sim_gen()}}.}
\usage{
id_plot_sims(sims, type = "RMSE")

id_show_trues(sims, type = "RMSE")
}
\arguments{
\item{sims}{A fitted \code{idealstan} object that has true data generated by \code{\link[=id_sim_gen]{id_sim_gen()}}}

\item{type}{Type of analysis of true versus fitted values, can be \code{'RMSE'}, \code{'Residuals'} or \code{'Coverage'}}
}
\value{
A \code{ggplot2} object showing RMSE, residuals, or coverage for simulated parameters
compared to their true values.
}
\description{
This function plots the results from a simulation generated by \code{\link[=id_sim_gen]{id_sim_gen()}}.
}
\examples{
\donttest{
sim <- id_sim_gen()
est <- id_estimate(sim, model_type=1, fixtype='vb_full',
                   use_method="pathfinder", nchains=2, ncores=2)
id_plot_sims(est)               # RMSE (default)
id_plot_sims(est, type='Residuals')
id_plot_sims(est, type='Coverage')
}
}
