% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{id_post_pred,idealstan-method}
\alias{id_post_pred,idealstan-method}
\title{Posterior Prediction for \code{idealstan} objects}
\usage{
\S4method{id_post_pred}{idealstan}(
  object,
  newdata = NULL,
  draws = 100,
  output = "observed",
  type = "predict",
  covar = "person",
  sample_scores = NULL,
  item_subset = NULL,
  pred_outcome = NULL,
  use_cores = 1,
  use_chain = NULL,
  skip_cov = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted \code{idealstan} object}

\item{newdata}{Optional: pass a data frame that must have all of the predictors that
were given to the id_make function. Used to generate predictions from person or item
covariates on to items.}

\item{draws}{The number of draws to use from the total number of posterior draws (default is 100). Set to "all" to use all draws in the chains.
For reproducibility, you can also pass a vector of specific draws to use.}

\item{output}{If the model has an unbounded outcome (Poisson, continuous, etc.), then
specify whether to show the \code{'observed'} data (the default) or the binary
output \code{'missing'} showing whether an observation was predicted as missing or not}

\item{type}{Whether to produce posterior predictive values (\code{'predict'}, the default),
the posterior expected (average) values (\code{'epred'}),
or log-likelihood values (\code{'log_lik'}). See the How to Evaluate Models vignette for more info.}

\item{covar}{What kind of covariates to include as part of the prediction -- either
"person" (the default) or "items" if you included predictors for item discriminations.}

\item{sample_scores}{In addition to reducing the number of posterior draws used to
calculate the posterior predictive distribution, which will reduce computational overhead.
Only available for calculating predictive distributions, not log-likelihood values.}

\item{item_subset}{Whether to calculate marginal effects for only a subset of
items. Should be item IDs that match the \code{item_id} column passed to the \code{id_make}
function.}

\item{pred_outcome}{In the case of ordinal responses, the number of the category to
predict. Defaults to top category.}

\item{use_cores}{Number of cores to use for multicore parallel processing with
the base R \code{parallel} package}

\item{use_chain}{ID of MCMC chain to use rather than all chains (the default).}

\item{skip_cov}{Whether to skip adding in hierarchical person-level covariates. Defaults to FALSE.}

\item{...}{Any other arguments passed on to posterior_predict (currently none available)}
}
\value{
An object of class \code{id_pred_obj} (a list of matrices of posterior-predicted draws, one
per model type). If \code{type="log_lik"}, a draws-by-observations \code{matrix} of class
\code{log_lik} with a \code{chain_order} attribute for use with \code{\link{derive_chain}}.
}
\description{
This function will draw from the posterior distribution, whether in terms of the outcome (prediction)
or to produce the log-likelihood values.

This function can also produce either distribution of the
outcomes (i.e., predictions) or the log-likelihood values of the posterior (set option
\code{type} to \code{'log_lik'}.
For more information, see the package vignette How to Evaluate Models.

You can then use functions such as
\code{\link[=id_plot_ppc]{id_plot_ppc()}} to see how well the model does returning the correct number of categories
in the score/vote matrix.
Also see \code{help("posterior_predict", package = "rstanarm")}
}
\examples{
\donttest{
sim <- id_sim_gen()
est <- id_estimate(sim, model_type=1, fixtype='vb_full',
                   use_method="pathfinder", nchains=2, ncores=2)
# posterior predictive distribution (100 draws)
ppc <- id_post_pred(est, draws=50)
# log-likelihood matrix (for loo)
ll <- id_post_pred(est, type='log_lik')
}
}
