% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{id_rebuild_mpi}
\alias{id_rebuild_mpi}
\title{Reconstitute an idealstan object after an MPI/cluster run}
\usage{
id_rebuild_mpi(file_loc = NULL, csv_name = NULL)
}
\arguments{
\item{file_loc}{A string with the location of the original files exported by 
\code{id_estimate}}

\item{csv_name}{A vector of character names of CSV files with posterior estimates from 
\code{cmdstan}. Should be located in the same place as \code{file_loc}.}

\item{object}{A fitted \code{idealstan} object (see \code{\link{id_estimate}})}
}
\description{
This convenience function takes as input a file location storing the results of a 
MPI/cluster run.
}
\details{
Given the CSV output from cmdstan and the original files exported 
from the \code{id_estimate} function, this function will create an \code{idealstan}
object which can be further analyzed with other \code{idealstan} package helper
functions.
}
