% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\docType{class}
\name{idealdata-class}
\alias{idealdata-class}
\title{Data and Identification for \code{id_estimate}}
\description{
\code{idealdata} objects contain the relevant legislator/bill (person/item) matrix of data along with slots containing information
about the kind of identification used in the estimation.
}
\seealso{
\code{\link{id_make}} to create an \code{idealdata} object suitable for estimation with \code{id_estimate}.
}
