\name{idealstan-package}
\alias{idealstan-package}
\alias{idealstan}
\docType{package}
\title{
\packageTitle{idealstan}
}
\description{
\packageDescription{idealstan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{idealstan}
\packageIndices{idealstan}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{idealstan}

Maintainer: \packageMaintainer{idealstan}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
