% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealstan.R
\docType{package}
\name{idealstan}
\alias{idealstan}
\alias{idealstan-package}
\title{\code{idealstan} package}
\description{
R Interface to Stan for Item-Response Theory Ideal Point Models
}
\details{
See the README on
\href{https://github.com/saudiwin/idealstan/blob/master/README.md}{GitHub}
}
\references{
\enumerate{
   \item Kubinec, Robert. Generalized Ideal Point Models for Time-Varying and Missing-Data Inference. Working Paper.
   \item Clinton, J., Jackman, S., & Rivers, D. (2004). The Statistical Analysis of Roll Call Data. \emph{The American Political Science Review}, 98(2), 355-370. doi:10.1017/S0003055404001194
   \item Bafumi, J., Gelman, A., Park, D., & Kaplan, N. (2005). Practical Issues in Implementing and Understanding Bayesian Ideal Point Estimation. \emph{Political Analysis}, 13(2), 171-187. doi:10.1093/pan/mpi010
}
}
