% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{launch_shinystan,idealstan-method}
\alias{launch_shinystan,idealstan-method}
\title{Function to Launch Shinystan with an \code{idealstan} Object}
\usage{
\S4method{launch_shinystan}{idealstan}(
  object,
  pars = c("L_full", "sigma_reg_full", "sigma_abs_free", "A_int_free", "B_int_free",
    "steps_votes", "steps_votes_grm"),
  ...
)
}
\arguments{
\item{object}{A fitted \code{idealstan} object}

\item{pars}{A character vector of parameters to select from the underlying \code{rstan} model object}

\item{...}{Other parameters passed on to \code{\link[shinystan]{shinystan}}}
}
\value{
No return value; called for side effects (launches the Shinystan interactive diagnostics application).
}
\description{
This wrapper will pull the \code{rstan} samples out of a fitted \code{idealstan} model and then launch
\code{\link[shinystan]{launch_shinystan}}. This function is useful for examining convergence statistics of the
underlying MCMC sampling.
}
\examples{
\donttest{
data('senate114')
senate114$cast_code <- ifelse(senate114$cast_code=="Absent", NA,
                              as.integer(senate114$cast_code) - 1L)
sen_data <- id_make(senate114, outcome_disc='cast_code',
                    person_id='bioname', item_id='rollnumber',
                    group_id='party_code')
sen_est <- id_estimate(sen_data, model_type=1, use_method="pathfinder", fixtype='vb_full', ncores=4)

# will launch interactive browser
if(interactive()) {
  launch_shinystan(sen_est)
}

}
}
\seealso{
\code{\link[shinystan]{shinystan}}
}
