% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{launch_shinystan}
\alias{launch_shinystan}
\title{Generic Method to Use \code{shinystan} with \code{idealstan}}
\usage{
launch_shinystan(object, ...)
}
\arguments{
\item{object}{A fitted \code{idealstan} object.}

\item{...}{Other arguments passed on to underlying function}
}
\value{
No return value; called for side effects (launches the Shinystan interactive diagnostics application).
}
\description{
A generic function for launching \code{\link[shinystan]{launch_shinystan}}.
}
