% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\docType{methods}
\name{plot_model,idealstan-method}
\alias{plot_model,idealstan-method}
\title{The base plotting function. Default plot shows legislator ideal points with bills as equiprobability lines 
(also called trace contour plots and cutting lines).}
\usage{
\S4method{plot_model}{idealstan}(object, plot_type = "legislators", ...)
}
\description{
The base plotting function. Default plot shows legislator ideal points with bills as equiprobability lines 
(also called trace contour plots and cutting lines).
}
