% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Generic function for drawing from the posterior predictive distribution}
\usage{
posterior_predict(object, ...)
}
\arguments{
\item{object}{A fitted \code{idealstan} object}

\item{...}{All other parameters passed on to the underlying function.}
}
\value{
\code{posterior_predict} methods should return a \eqn{D} by \eqn{N}
  matrix, where \eqn{D} is the number of draws from the posterior predictive
  distribution and \eqn{N} is the number of data points being predicted per
  draw.
}
\description{
This generic is provided for compability with \code{\link[rstan]{stan}}.
}
