% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealstan.R
\name{release_questions}
\alias{release_questions}
\title{Function that provides additional check questions for package release}
\usage{
release_questions()
}
\description{
Function that provides additional check questions for package release
}
