% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_doc.R
\docType{data}
\name{senate114_fit}
\alias{senate114_fit}
\title{Fitted \code{idealstan} model for 114th Senate Data}
\format{A fitted \code{idealstan} object}
\usage{
senate114_fit
}
\description{
This R object is a fitted \code{idealstan} object that can be used for testing and illustratory purposes.
}
\keyword{datasets}
