% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{stan_trace,idealstan-method}
\alias{stan_trace,idealstan-method}
\title{Plot the MCMC posterior draws by chain}
\usage{
\S4method{stan_trace}{idealstan}(object, par = "L_full[1]", ...)
}
\arguments{
\item{object}{A fitted \code{idealstan} model}

\item{par}{The character string  name of a parameter in the model}

\item{...}{Other options passed on to \code{\link[bayesplot]{mcmc_trace}}}
}
\value{
A \code{ggplot2} object (from \code{\link[bayesplot]{mcmc_trace}}) showing MCMC trace plots
for the selected parameter.
}
\description{
This function allows you to produce trace plots for assessing the quality
and convergence of MCMC chains.
}
\details{
To use this function, you must pass a fitted \code{idealstan} object
along with the name of a parameter in the model. To determine these
parameter names, use the \code{summary} function or obtain the data
from a plot by passing the \code{return_data=TRUE} option to
\code{id_plot_persons} or \code{id_plot_persons_dyn} to find the
name of the parameter in the Stan model.

This function is a simple wrapper around \code{\link[bayesplot]{mcmc_trace}}.
Please refer to that function's documentation for further options.
}
\examples{
\donttest{
data('senate114')
senate114$cast_code <- ifelse(senate114$cast_code=="Absent", NA,
                              as.integer(senate114$cast_code) - 1L)
sen_data <- id_make(senate114, outcome_disc='cast_code',
                    person_id='bioname', item_id='rollnumber',
                    group_id='party_code')
sen_est <- id_estimate(sen_data, model_type=1,
            fixtype='vb_full', use_method="pathfinder", 
            ncores=4)
stan_trace(sen_est, par='L_full[1]')
}
}
