% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{.prepare_legis_data}
\alias{.prepare_legis_data}
\title{Helper function for preparing person ideal point plot data}
\usage{
.prepare_legis_data(object, high_limit = NULL, low_limit = NULL,
  aggregate = TRUE)
}
\description{
Helper function for preparing person ideal point plot data
}
