% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{estimate_ideal}
\alias{estimate_ideal}
\title{Estimate an idealstan model using an idealdata object.}
\usage{
estimate_ideal(idealdata = NULL, use_subset = FALSE, sample_it = FALSE,
  subset_party = NULL, subset_legis = NULL, sample_size = 20,
  nchains = 4, niters = 2000, use_vb = FALSE, nfix = 10,
  fixtype = "vb", warmup = floor(niters/2), ncores = NULL,
  modeltype = "ratingscale_absence_inflate", ...)
}
\description{
Estimate an idealstan model using an idealdata object.
}

