% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{estimate_ideal}
\alias{estimate_ideal}
\title{Estimate an idealstan model using an idealdata object.}
\usage{
estimate_ideal(idealdata = NULL, model_type = 2, use_subset = FALSE,
  sample_it = FALSE, subset_party = NULL, subset_legis = NULL,
  sample_size = 20, nchains = 4, niters = 2000, use_vb = FALSE,
  nfix = 1, restrict_params = "bill", pin_vals = NULL,
  restrict_ind_high = NULL, restrict_ind_low = NULL,
  restrict_type = "constrain_oneway", fixtype = "vb",
  warmup = floor(niters/2), ncores = NULL, auto_id = FALSE, ...)
}
\description{
Estimate an idealstan model using an idealdata object.
}
