% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{id_extract}
\alias{id_extract}
\title{Generic Method for Extracing Posterior Samples}
\usage{
id_extract(object, ...)
}
\arguments{
\item{object}{A fitted \code{idealstan} object}

\item{...}{Other arguments passed on to underlying functions}
}
