% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{id_make}
\alias{id_make}
\title{Create data to run IRT model}
\usage{
id_make(score_data = NULL, outcome = "outcome",
  person_id = "person_id", item_id = "item_id", time_id = "time_id",
  group_id = "group_id", simul_data = NULL, person_cov = NULL,
  group_cov = NULL, varying_group_id = NULL, item_cov = NULL,
  miss_cov = NULL, miss_exog_val = NULL, item_cov_miss = NULL,
  person_data = data_frame(), item_data = NULL, miss_val = NA,
  high_val = 3L, low_val = 1L, middle_val = 2L, ordinal = FALSE,
  continuous = FALSE, time = NULL, outcome_label_type = "votes",
  exclude_level = NULL, inflate = TRUE, simulation = FALSE,
  include_pres = FALSE)
}
\arguments{
\item{score_data}{A data frame in long form, i.e., one row in the data for each 
measured score or vote in the data.}

\item{outcome}{Column name of the outcome in \code{score_data}, default is \code{"outcome"}}

\item{person_id}{Column name of the person/legislator ID index in \code{score_data}, 
default is \code{'person_id'}}

\item{item_id}{Column name of the item/bill ID index in \code{score_data}, 
default is \code{'item_id'}}

\item{time_id}{Column name of the time values in \code{score_data}: 
optional, default is \code{'time_id'}}

\item{group_id}{Optional column name of a person/legislator group IDs (i.e., parties) in \code{score_data}. 
Optional, default is \code{'group_id'}}

\item{simul_data}{Optionally, data that has been generated by the \code{\link{id_sim_gen}} function.}

\item{person_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the person/legislator ideal points}

\item{group_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the person/legislator ideal points at the group level}

\item{item_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the 
discrimination parameters for the regular model}

\item{miss_exog_val}{An integer or string indicating values that should be treated as exogenous missing values in the score (rollcall) data. This value must be included in the list of missing values \code{miss_val}.}

\item{item_cov_miss}{A one-sided formula that specifies the covariates
in the dataset that will be used to hierarchically model the discrimination parameters for the
missing data model}

\item{person_data}{An optional data frame of additional information about the persons (legislators) used for plotting. 
Must have a
column of the same name as \code{person_id} to match back to the long data.
Suggested other columns include a \code{person.names} column containing names of
 persons (legislators) and optionally a \code{group} column with names of any groupings of 
 the persons (legislators), such as parties or blocs (used for visualization).}

\item{item_data}{An optional data frame of item/bill labels and other information used for visualization. Should include \code{item.names} column 
containing item/bill labels; other columns are optional.}

\item{miss_val}{The value (numeric or character) that indicate missing data/absences in the data. If there are multiple possible values, 
pass along a numeric or character vector of all such values. If missing data is coded as \code{NA}, simply leave this parameter at the default, \code{NA}.}

\item{high_val}{The value (numeric or character) that indicate the highest ordinal outcome possible, such as yes in a vote dataset or correct in a test examination.
If there are multiple possible values, 
pass along a numeric or character vector of all such values.}

\item{low_val}{The value (numeric or character) that indicates the lowest ordinal outcome possible, such as no votes in a vote dataset or incorrect in a test examination.
If there are multiple possible values, 
pass along a numeric or character vector of all such values.}

\item{middle_val}{The value (numeric or character) that indicate values between the lowest and highest categories, such as abstention in voting data or "Neither Agree nor Disagree" in likert scales.
If there are multiple possible values, 
pass along a numeric or character vector of all such values in correct order (lower to higher values).
If there are no middle values (binary outcome), set to \code{NULL}.}

\item{ordinal}{Whether or not the data contain ordinal responses. If \code{TRUE}, middle values/abstentions are used as a middle category in constructing the outcome.
Otherwise the response is assumed to be binary (yes/no) or (correct/incorrect).
@param continuous Whether or not the outcome/response is continuous. If it is, \code{miss_val} is recoded as the maximum of the outcome + 1.}

\item{exclude_level}{A vector of any values that should be treated as \code{NA} in the response matrix. 
Unlike the \code{middle_val} parameter, these values will be dropped from the data before estimation rather than modeled explicitly.}

\item{inflate}{If \code{TRUE}, the score matrix is set up to enable modeling of 
missing data/absences (\code{miss_val}) as an inflation model in \code{\link{id_estimate}}}

\item{simulation}{If \code{TRUE}, simulated values are saved in the \code{idealdata} object for later plotting with the \code{\link{id_plot_sims}} function}

\item{include_pres}{If \code{FALSE} and \code{score_data} is a \code{rollcall} object, drop the first row which often represents tiebreaker votes cast by the Vice President in the US Senate.}
}
\value{
A \code{idealdata} object that can then be used in the \code{\link{id_estimate}} function to fit a model.
}
\description{
To run an IRT model using \code{idealstan}, you must first process your data using the \code{id_make} function.
}
\examples{
# You can either use a pscl rollcall object or a vote/score matrix 
# where persons/legislators are in the rows
# and items/bills are in the columns

library(dplyr)

# First, using a rollcall object with the 114th Senate's rollcall votes:

data('senate114')

to_idealstan <-   id_make(score_data = senate114,
ordinal = FALSE,
include_pres=FALSE)

}
