% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{id_make}
\alias{id_make}
\title{Create data to run IRT model}
\usage{
id_make(score_data = NULL, outcome = "outcome", person_id = "person_id",
  item_id = "item_id", time_id = "time_id", group_id = "group_id",
  simul_data = NULL, person_cov = NULL, group_cov = NULL,
  item_cov = NULL, miss_cov = NULL, item_cov_miss = NULL, miss_val = NA,
  high_val = 3L, low_val = 1L, middle_val = 2L, ordinal = FALSE,
  unbounded = FALSE, exclude_level = NULL, inflate = TRUE,
  simulation = FALSE)
}
\arguments{
\item{score_data}{A data frame in long form, i.e., one row in the data for each 
measured score or vote in the data or a \code{rollcall} data object from package \code{pscl}.}

\item{outcome}{Column name of the outcome in \code{score_data}, default is \code{"outcome"}}

\item{person_id}{Column name of the person/legislator ID index in \code{score_data}, 
default is \code{'person_id'}. Should be integer, character or factor.}

\item{item_id}{Column name of the item/bill ID index in \code{score_data}, 
default is \code{'item_id'}.  Should be integer, character or factor.}

\item{time_id}{Column name of the time values in \code{score_data}: 
optional, default is \code{'time_id'}. Should be a date or date-time class, but can be an integer
(i.e., years in whole numbers).}

\item{group_id}{Optional column name of a person/legislator group IDs (i.e., parties) in \code{score_data}. 
Optional, default is \code{'group_id'}. Should be integer, character or factor.}

\item{simul_data}{Optionally, data that has been generated by the \code{\link{id_sim_gen}} function.}

\item{person_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the person/legislator ideal points}

\item{group_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the person/legislator ideal points 
at the group level. Use this in place of \code{person_cov} if you intend to run a group-level model.}

\item{item_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the 
item/bill discrimination parameters for the regular model}

\item{item_cov_miss}{A one-sided formula that specifies the covariates
in the dataset that will be used to hierarchically model the item/bill discrimination parameters for the
missing data model.}

\item{miss_val}{The value (numeric or character) that indicate missing data/absences in the data. If 
\code{inflate==TRUE}, these values will be recoded to a separate category to use in the inflation model.
If there are multiple possible values, 
 pass along a numeric or character vector of all such values. If missing data is coded as \code{NA}, 
 simply leave this parameter at the default, \code{NA}.}

\item{high_val}{The value (numeric or character) that indicate the highest ordinal outcome possible, 
such as yes in a vote dataset or correct in a test examination.
 If there are multiple possible values, 
 pass along a numeric or character vector of all such values.}

\item{low_val}{The value (numeric or character) that indicates the lowest ordinal outcome possible, 
such as no votes in a vote dataset or incorrect in a test examination.
 If there are multiple possible values, 
 pass along a numeric or character vector of all such values.}

\item{middle_val}{The value (numeric or character) that indicate values 
between the lowest and highest categories, such as abstention in voting data or "Neither Agree nor Disagree" in likert scales.
 If there are multiple possible values, 
 pass along a numeric or character vector of all such values in correct order (lower to higher values).
 If there are no middle values (binary outcome), set to \code{NULL}.}

\item{ordinal}{Whether or not the data contain ordinal responses. If \code{TRUE},
middle values/abstentions are used as a middle category in constructing the outcome.
Otherwise the response is assumed to be binary (yes/no).}

\item{unbounded}{Whether or not the outcome/response is unbounded (i.e., continuous or
Poisson). If it is, \code{miss_val} 
is recoded as the maximum of the outcome + 1.}

\item{exclude_level}{A vector of any values that should be treated as \code{NA} in the response matrix. 
Unlike the \code{miss_val} parameter, these values will be dropped from the data before 
estimation rather than modeled explicitly.}

\item{inflate}{If \code{TRUE}, the score matrix is set up to enable modeling of 
missing data/absences (\code{miss_val}) as an inflation model in \code{\link{id_estimate}}}

\item{simulation}{If \code{TRUE}, simulated values are saved in the \code{idealdata} object for later plotting with the \code{\link{id_plot_sims}} function}
}
\value{
A \code{idealdata} object that can then be used in the \code{\link{id_estimate}} function 
to fit a model.
}
\description{
To run an IRT model using \code{idealstan}, you must first process your data using the \code{id_make} 
function.
}
\details{
This function can be used either on a \code{rollcall} data object from package
\code{pscl} or on a long data frame where one row equals one item-person (bill-legislator)
observation. If a \code{rollcall} object is input, the \code{rollcall} data is converted
to a long data frame.
If passing a long data frame, you should specify the names of the 
columns containing the IDs for persons, items and 
groups (groups are IDs that may have multiple observations per ID, such as political parties or
classes) to the \code{id_make} function. 
The only required columns are the item/bill ID and the person/legislator ID along with an 
outcome column. 
To run a time-varying model, you need to include a column with dates (or integers). To run an 
inflated model for missing data, be sure to specify the \code{miss_val} option to the value that should
be modeled as missing and also set \code{inflate} to \code{TRUE}. Otherwise the \code{id_estimate}
function will not have data in the correct form to run an inflated model.
If the outcome is unbounded, simply set \code{unbounded} to \code{TRUE}. You can ignore the
options that specify which values should be \code{high_val} or \code{low_val}, and all 
missing values (\code{NA}) will be recoded to a specific value out of the range of the outcome to use
for inflation.
}
\examples{
# You can either use a pscl rollcall object or a vote/score matrix 
# where persons/legislators are in the rows
# and items/bills are in the columns

library(dplyr)

# First, using a rollcall object with the 114th Senate's rollcall votes:

data('senate114')

to_idealstan <-   id_make(score_data = senate114,
ordinal = FALSE)

}
