% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{id_make}
\alias{id_make}
\title{Create data to run IRT model}
\usage{
id_make(
  score_data = NULL,
  outcome_disc = "outcome_disc",
  outcome_cont = "outcome_cont",
  person_id = "person_id",
  item_id = "item_id",
  time_id = "time_id",
  group_id = "group_id",
  model_id = "model_id",
  ordered_id = "ordered_id",
  ignore_id = "ignore_id",
  simul_data = NULL,
  person_cov = NULL,
  item_cov = NULL,
  item_cov_miss = NULL,
  remove_cov_int = FALSE,
  unbounded = FALSE,
  exclude_level = NA,
  simulation = FALSE
)
}
\arguments{
\item{score_data}{A data frame in long form, i.e., one row in the data for each
measured score or vote in the data or a \code{rollcall} data object from package \code{pscl}.}

\item{outcome_disc}{Column name of the outcome with discrete values in \code{score_data}, default is \code{"outcome_disc"}}

\item{outcome_cont}{Column name of the outcome with discrete values in \code{score_data}, default is \code{"outcome_disc"}}

\item{person_id}{Column name of the person/legislator ID index in \code{score_data},
default is \code{'person_id'}. Should be integer, character or factor.}

\item{item_id}{Column name of the item/bill ID index in \code{score_data},
default is \code{'item_id'}.  Should be integer, character or factor.}

\item{time_id}{Column name of the time values in \code{score_data}:
optional, default is \code{'time_id'}. Should be a date or date-time class, but can be an integer
(i.e., years in whole numbers).}

\item{group_id}{Optional column name of a person/legislator group IDs (i.e., parties) in \code{score_data}.
Optional, default is \code{'group_id'}. Should be integer, character or factor.}

\item{model_id}{Column name of the model/response types in the data.
Default is \code{"model_id"}. Only necessary if a model with multiple
response types (i.e., binary + continuous outcomes). Must be a
column with a series
of integers matching the model types in \code{\link[=id_estimate]{id_estimate()}}
showing which row of the data matches which outcome.}

\item{ordered_id}{Column name of the variable showing the count of categories for
ordinal/categorical items (must be at least 3 categories)}

\item{ignore_id}{Optional column for identifying observations that should not be
modeled (i.e., not just treated as missing, rather removed during estimation). Should
be a binary vector (0 for remove and 1 for include). Useful for time-varying models where
persons may not be present during particular periods and missing data is ignorable.}

\item{simul_data}{Optionally, data that has been generated by the \code{\link[=id_sim_gen]{id_sim_gen()}} function.}

\item{person_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the person/legislator ideal points}

\item{item_cov}{A one-sided formula that specifies the covariates
in \code{score_data} that will be used to hierarchically model the
item/bill discrimination parameters for the regular model}

\item{item_cov_miss}{A one-sided formula that specifies the covariates
in the dataset that will be used to hierarchically model the item/bill discrimination parameters for the
missing data model.}

\item{remove_cov_int}{Whether to remove constituent terms from hierarchical covariates that
interact covariates with IDs like \code{person_id} or \code{item_id}. Set to \code{TRUE} if
including these constituent terms would cause multi-collinearity with other terms in the model
(such as running a group-level model with a group-level interaction or a person-level model
with a person-level interaction).}

\item{unbounded}{Whether or not the outcome/response is unbounded (i.e., continuous or
Poisson). If it is, missing value
is recoded as the maximum of the outcome + 1.}

\item{exclude_level}{A vector of any values that should be treated as \code{NA} in the response matrix.
Unlike missing values, these values will be dropped from the data before
estimation rather than modeled explicitly.}

\item{simulation}{If \code{TRUE}, simulated values are saved in the \code{idealdata} object for
later plotting with the \code{\link[=id_plot_sims]{id_plot_sims()}} function}
}
\value{
A \code{idealdata} object that can then be used in the \code{\link[=id_estimate]{id_estimate()}} function
to fit a model.
}
\description{
To run an IRT model using \code{idealstan}, you must first process your data using the \code{id_make}
function.
}
\details{
This function accepts a long data frame where one row equals one item-person (bill-legislator)
observation with associated continuous or discrete outcomes/responses.
You either need to include columns with specific names as required by the \code{id_make}
function such as \code{person_id} for person IDs and \code{item_id} for item IDs or
specify the names of the
columns containing the IDs to the \code{id_make} function for each column name (see examples).
The only required columns are the item/bill ID and the person/legislator ID along with an
outcome column, \code{outcome_disc} for discrete variables and \code{outcome_cont} for
continuous variables. If both columns are included, then any value can be included for
\code{outcome_disc} if there are values for \code{outcome_cont} and vice versa.

If items of multiple types are included, a column \code{model_id} must be included with
the model type (see \code{id_estimate} function documentation for list of model IDs)
for the response distribution, such as
1 for binary non-inflated, etc. If an ordinal outcome is included, an additional column
\code{ordered_id} must be included that has the total count of categories for that
ordinal variable (i.e., 3 for 3 categories).

For discrete data, it is recommended to include a numeric variable that starts at 0, such
as values of 0 and 1 for binary data and 0,1,2 for ordinal/categorical data.
For continuous (unbounded) data, it is recommended to standardize the outcome to improve
model convergence and fit.

Missing data should be passed as \code{NA} values in either
\code{outcome_disc} or \code{outcome_cont} and will be processed internally.
}
\section{Time-Varying Models}{


To run a time-varying model, you need to include the name of a column with dates (or integers) that is passed
to the \code{time_id} option.
}

\section{Continuous Outcomes}{


If the outcome is continuous, you need to pass a dataframe with one column named
"outcome_disc" or pass the name of the column with the continuous data to the \code{outcome_disc}
argument.
}

\section{Hierarchical Covariates}{


Covariates can be fit on the person-level ideal point parameters as well as
item discrimination parameters for either the inflated (missing) or non-inflated (observed)
models. These covariates must be columns that were included with the data fed to the
\code{\link[=id_make]{id_make()}} function. The covariate relationships are specified as
one-sided formulas, i.e. \code{~cov1 + cov2 + cov1*cov2}. To interact covariates with the
person-level ideal points you can use \code{~cov1 + person_id + cov1*person_id} and for
group-level ideal poins you can use \code{~cov1 + group_id + cov1*group_id} where
\code{group_id} or \code{person_id} is the same name as the name of the column
for these options that you passed to \code{id_make} (i.e., the names of the columns
in the original data). If you are also going to model these intercepts--i.e. you are
interacting the covariate with \code{person_id} and the model is estimating ideal points
at the person level--then set \code{remove_cov_int} to TRUE to avoid multicollinearity with the
ideal point intercepts.
}

\examples{
# You can either use a pscl rollcall object or a vote/score matrix 
# where persons/legislators are in the rows
# and items/bills are in the columns

library(dplyr)

# First, using a rollcall object with the 114th Senate's rollcall votes:

data('senate114')

to_idealstan <-   id_make(score_data = senate114,
               outcome_disc = 'cast_code',
               person_id = 'bioname',
               item_id = 'rollnumber',
               group_id= 'party_code',
               time_id='date')

}
