% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{id_me}
\alias{id_me}
\title{Calculate ideal point marginal effects}
\usage{
id_me(object, ...)
}
\arguments{
\item{object}{A fitted \code{idealstan} model}

\item{...}{Other values passed on to methods.}
}
\value{
Returns a tibble that has one row per posterior draw per item-specific
marginal effect in the scale of the outcome.
}
\description{
This function allows you to calculate ideal point marginal effects for a
given person-level hierarchical covariate.
}
\details{
This function will calculate item-level ideal point marginal effects
for a given covariate that was passed to the \code{id_make} function using the
\code{person_cov} option. The function will iterate over all items in the model
and use numerical differentiation to calculate responses in the scale of the
outcome for each item. Note: if the covariate is binary (i.e., only has two values),
then the function will calculate the difference between these two values instead of
using numerical differentation.
}
