% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{id_plot_ppc,idealstan-method}
\alias{id_plot_ppc,idealstan-method}
\title{Plot Posterior Predictive Distribution for \code{idealstan} Objects}
\usage{
\S4method{id_plot_ppc}{idealstan}(object, ppc_pred = NULL, group = NULL, item = NULL, ...)
}
\arguments{
\item{object}{A fitted idealstan object}

\item{ppc_pred}{The output of the \code{\link{id_post_pred}} function on a fitted idealstan object}

\item{group}{A character vector of the person or group IDs 
over which to subset the predictive distribution}

\item{...}{Other arguments passed on to \code{\link[bayesplot]{ppc_bars}}}
}
\description{
This function is the actual method for generating posterior distributions 
from a fitted \code{idealstan} model.
}
\details{

}
