% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{id_plot_ppc,idealstan-method}
\alias{id_plot_ppc,idealstan-method}
\title{Plot Posterior Predictive Distribution for \code{idealstan} Objects}
\usage{
\S4method{id_plot_ppc}{idealstan}(
  object,
  ppc_pred = NULL,
  group = NULL,
  item = NULL,
  combine_item = TRUE,
  type = NULL,
  only_observed = NULL,
  which_mod = NULL,
  prompt_plot = TRUE,
  ...
)
}
\arguments{
\item{object}{A fitted idealstan object}

\item{ppc_pred}{The output of the \code{\link{id_post_pred}} function on a fitted idealstan object}

\item{group}{A character vector of the person or group IDs 
over which to subset the predictive distribution}

\item{item}{A character vector of the item IDs
over which to subset the predictive distribution}

\item{combine_item}{Whether to combine all items together (TRUE) or create one plot for each item (FALSE)}

\item{type}{Whether to plot "continuous" or "discrete" responses}

\item{which_mod}{If you are producing one plot aggregating data across multiple items and 
you have different item distributions,
then you need to specify the item type number to plot (see function documentation in
\code{\link{id_estimate}}).}

\item{prompt_plot}{Whether to expect a user prompt for each plot if multiple plots 
are produced (defaults to TRUE)
If NULL (default), will use the type specified in the data. 
However, if both continuous and discrete items are present, will
throw an error if NULL.}

\item{...}{Other arguments passed on to \code{\link[bayesplot]{ppc_bars}}}
}
\description{
This function is the actual method for generating posterior distributions 
from a fitted \code{idealstan} model.
}
\details{

}
