% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\docType{methods}
\name{id_plot_ppc,idealstan-method}
\alias{id_plot_ppc,idealstan-method}
\title{Plot Posterior Predictive Distribution for \code{idealstan} Objects}
\usage{
\S4method{id_plot_ppc}{idealstan}(object, ppc_pred = NULL, group = NULL,
  item = NULL, ...)
}
\arguments{
\item{object}{A fitted idealstan object}

\item{ppc_pred}{The output of the \code{\link{id_post_pred}} function on a fitted idealstan object}

\item{group}{A character vector of the person or group IDs 
over which to subset the predictive distribution}

\item{item}{A character vector of item IDs to calculate their model predictions}

\item{...}{Other arguments passed on to \code{\link[bayesplot]{ppc_bars}}}
}
\description{
This function is a wrapper around \code{\link[bayesplot]{ppc_bars}} that plots the posterior predictive distribution
derived from \code{\link{id_post_pred}} against the original data. You can also specify a legislator/person or
bill/item by specifying the ID of each in the original data as a character vector. 
Only persons or items can be specified,
not both.
}
