% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{id_plot_ppc}
\alias{id_plot_ppc}
\title{Plot Posterior Predictive Distribution for \code{idealstan} Objects}
\usage{
id_plot_ppc(object, ...)
}
\arguments{
\item{object}{A fitted idealstan object}

\item{...}{Other arguments passed on to \code{\link[bayesplot]{ppc_bars}}}

\item{ppc_pred}{The output of the \code{\link{id_post_pred}} function on a fitted idealstan object}

\item{group}{A character vector of the person or group IDs 
over which to subset the predictive distribution}

\item{item}{A character vector of item IDs to subset the posterior distribution}
}
