% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{id_plot_ppc}
\alias{id_plot_ppc}
\title{Plot Posterior Predictive Distribution for \code{idealstan} Objects}
\usage{
id_plot_ppc(object, ...)
}
\arguments{
\item{object}{A fitted \code{idealstan} object}

\item{...}{Other arguments passed on to \code{\link[bayesplot]{ppc_bars}}}
}
\description{
This function is the generic method for generating posterior distributions 
from a fitted \code{idealstan} model. Functions are documented in the 
actual method (\code{\linkS4method{id_plot_ppc}}).
}
