% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{id_plot_rhats}
\alias{id_plot_rhats}
\title{Plotting Function to Display Rhat Distribution}
\usage{
id_plot_rhats(obj)
}
\arguments{
\item{obj}{A fitted \code{idealstan} object.}
}
\description{
This plotting function displays a histogram of the Rhat values of all parameters in an \code{idealstan} model.
}
