% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{id_plot_sims}
\alias{id_plot_sims}
\title{This function plots the results from a simulation generated by one of the *_simulate functions}
\usage{
id_plot_sims(sims, type = "RMSE")
}
\description{
This function plots the results from a simulation generated by one of the *_simulate functions
}
