% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{id_post_pred,idealstan-method}
\alias{id_post_pred,idealstan-method}
\title{Posterior Prediction for \code{idealstan} objects}
\usage{
\S4method{id_post_pred}{idealstan}(
  object,
  draws = 100,
  output = "observed",
  type = "predict",
  covar = "person",
  sample_scores = NULL,
  item_subset = NULL,
  use_cores = 1,
  use_chain = NULL,
  newdata = NULL,
  ...
)
}
\arguments{
\item{object}{A fitted \code{idealstan} object}

\item{draws}{The number of draws to use from the total number of posterior draws (default is 100).
For reproducibility, you can also pass a vector of specific draws to use.}

\item{output}{If the model has an unbounded outcome (Poisson, continuous, etc.), then
specify whether to show the \code{'observed'} data (the default) or the binary 
output \code{'missing'} showing whether an observation was predicted as missing or not}

\item{type}{Whether to produce posterior predictive values (\code{'predict'}, the default),
the posterior expected (average) values (\code{'epred'}),
or log-likelihood values (\code{'log_lik'}). See the How to Evaluate Models vignette for more info.}

\item{covar}{What kind of covariates to include as part of the prediction -- either
"person" (the default) or "items" if you included predictors for item discriminations.}

\item{sample_scores}{In addition to reducing the number of posterior draws used to 
calculate the posterior predictive distribution, which will reduce computational overhead.
Only available for calculating predictive distributions, not log-likelihood values.}

\item{item_subset}{Whether to calculate marginal effects for only a subset of 
items. Should be item IDs that match the \code{item_id} column passed to the \code{id_make}
 function.}

\item{use_cores}{Number of cores to use for multicore parallel processing with
the base R \code{parallel} package}

\item{newdata}{Optional: pass a data frame that must have all of the predictors that
were given to the id_make function. Used to generate predictions from person or item
covariates on to items.}

\item{...}{Any other arguments passed on to posterior_predict (currently none available)}
}
\description{
This function will draw from the posterior distribution, whether in terms of the outcome (prediction)
or to produce the log-likelihood values.  
 
 This function can also produce either distribution of the 
 outcomes (i.e., predictions) or the log-likelihood values of the posterior (set option 
 \code{type} to \code{'log_lik'}.
 For more information, see the package vignette How to Evaluate Models.
 
 You can then use functions such as 
 \code{\link{id_plot_ppc}} to see how well the model does returning the correct number of categories
 in the score/vote matrix. 
 Also see \code{help("posterior_predict", package = "rstanarm")}
}
