% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{id_sim_coverage}
\alias{id_sim_coverage}
\title{Function that computes how often the true value of the parameter is included within the
95/5 high posterior density interval}
\usage{
id_sim_coverage(obj, rep = 1, quantiles = c(0.95, 0.05))
}
\arguments{
\item{obj}{A fitted \code{idealstan} object with true data generated by \code{\link[=id_sim_gen]{id_sim_gen()}}}

\item{rep}{How many times the models were fitted on new data, currently can only be 1}

\item{quantiles}{What the quantile coverage of the high posterior density interval should be}
}
\description{
Function that computes how often the true value of the parameter is included within the
95/5 high posterior density interval
}
