% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{id_sim_gen}
\alias{id_sim_gen}
\title{A function designed to simulate absence-inflated data with either binary or ordinal outcomes}
\usage{
id_sim_gen(num_legis = 50, num_bills = 50, absence_discrim_sd = 1,
  absence_diff_mean = 0.5, reg_discrim_sd = 1, diff_sd = 1,
  ideal_pts_sd = 1, prior_type = "gaussian", ordinal = TRUE,
  ordinal_outcomes = 3, graded_response = FALSE, absence = TRUE)
}
\description{
A function designed to simulate absence-inflated data with either binary or ordinal outcomes
}
