% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{id_sim_resid}
\alias{id_sim_resid}
\title{Residual function for checking estimated samples compared to true simulation scores
Returns a data frame with residuals plus quantiles.}
\usage{
id_sim_resid(obj, rep = 1)
}
\arguments{
\item{obj}{A fitted \code{idealstan} object with true data from \code{\link{id_sim_gen}}}

\item{rep}{Over how many replicates to calculate residuals? Currently can only be 1}
}
\description{
Residual function for checking estimated samples compared to true simulation scores
Returns a data frame with residuals plus quantiles.
}
