% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{id_sim_rmse}
\alias{id_sim_rmse}
\title{RMSE function for calculating individual RMSE values compared to true simulation scores
Returns a data frame with RMSE plus quantiles.}
\usage{
id_sim_rmse(obj, rep = 1)
}
\arguments{
\item{rep}{Over how many replicates to calculate RMSE? Currently can only be 1}

\item{object}{A fitted \code{idealstan} object with true data from \code{\link{id_sim_gen}}}
}
\description{
RMSE function for calculating individual RMSE values compared to true simulation scores
Returns a data frame with RMSE plus quantiles.
}
