% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{id_sim_rmse}
\alias{id_sim_rmse}
\title{RMSE function for calculating individual RMSE values compared to true simulation scores
Returns a data frame with RMSE plus quantiles.}
\usage{
id_sim_rmse(obj, rep = 1)
}
\arguments{
\item{est_param}{A matrix of posterior draws of a parameter}

\item{true_param}{A matrix (one column) of true parameter values}
}
\description{
RMSE function for calculating individual RMSE values compared to true simulation scores
Returns a data frame with RMSE plus quantiles.
}
