% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_doc.R
\docType{package}
\name{idealstan-package}
\alias{idealstan}
\alias{idealstan-package}
\title{Estimate Bayesian IRT ideal point models with Stan}
\description{
Idealstan enables you to run IRT-based ideal point models with the Stan Bayesian estimation engine. In addition, it includes
models that allow for censoring of the data, such as that occurs when legislators/persons are absent, and for ordinal outcomes,
such as when legislators abstain.
}
\details{
To get started, first look at how to pre-process your data with the \code{\link{id_make}} function, and then how to run a model
with the \code{\link{id_estimate}} function. To report a bug, please go to \url{www.github.com/saudiwin/idealstan} and put in an
issue in my Github page. You can email me at \email{rmk7xy@virginia.edu}.
}
