% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idealstan.R
\docType{package}
\name{idealstan}
\alias{idealstan}
\alias{idealstan-package}
\alias{idealstan-package}
\title{\code{idealstan} package}
\description{
R Interface to Stan for Item-Response Theory Ideal Point Models
}
\details{
See the README on
\href{https://github.com/saudiwin/idealstan/blob/master/README.md}{GitHub}
}
