% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{launch_shinystan,idealstan-method}
\alias{launch_shinystan,idealstan-method}
\title{Function to Launch Shinystan with an \code{idealstan} Object}
\usage{
\S4method{launch_shinystan}{idealstan}(
  object,
  pars = c("L_full", "sigma_reg_full", "sigma_abs_free", "A_int_free", "B_int_free",
    "steps_votes", "steps_votes_grm"),
  ...
)
}
\arguments{
\item{object}{A fitted \code{idealstan} object}

\item{pars}{A character vector of parameters to select from the underlying \code{rstan} model object}

\item{...}{Other parameters passed on to \code{\link[shinystan]{shinystan}}}
}
\description{
This wrapper will pull the \code{rstan} samples out of a fitted \code{idealstan} model and then launch
\code{\link[shinystan]{launch_shinystan}}. This function is useful for examining convergence statistics of the
underlying MCMC sampling.
}
\seealso{
\code{\link[shinystan]{shinystan}}
}
