% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\docType{methods}
\name{launch_shinystan,idealstan-method}
\alias{launch_shinystan,idealstan-method}
\title{Function to Launch Shinystan with an \code{idealstan} Object}
\usage{
\S4method{launch_shinystan}{idealstan}(object, pars = c("L_free",
  "sigma_reg_free", "sigma_abs_free", "restrict_high", "restrict_low",
  "restrict_ord", "steps_votes", "steps_votes_grm"), ...)
}
\description{
This wrapper will pull the \code{rstan} samples out of a fitted \code{idealstan} model and then launch
\code{\link[shinystan]{launch_shinystan}}. This function is useful for examining convergence statistics of the 
underlying MCMC sampling.
}
\seealso{
\code{\link[shinystan]{shinystan}}
}
