% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{make_idealdata}
\alias{make_idealdata}
\title{This is the constructor fucnction for the idealdata object, which is necessary to estimate an idealstan model.}
\usage{
make_idealdata(vote_data = NULL, legis_data = NULL, bill_data = NULL,
  votes = NULL, abs_vote = NA)
}
\description{
This is the constructor fucnction for the idealdata object, which is necessary to estimate an idealstan model.
}

