% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstan_generics.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Generic function for drawing from the posterior predictive distribution}
\usage{
posterior_predict(object, ...)
}
\value{
\code{posterior_predict} methods should return a \eqn{D} by \eqn{N}
  matrix, where \eqn{D} is the number of draws from the posterior predictive
  distribution and \eqn{N} is the number of data points being predicted per
  draw.
}
\description{
Draw from the posterior predictive distribution of the outcome. See
\code{\link[rstanarm]{posterior_predict.stanreg}} in the
\pkg{\link[rstanarm]{rstanarm}} package for an example.
}
\examples{
# Example using rstanarm package:
# posterior_predict method for 'stanreg' objects
\donttest{
if (require("rstanarm")) {
  fit <- stan_glm(mpg ~ wt + am, data = mtcars)
  yrep <- posterior_predict(fit)
  all.equal(ncol(yrep), nobs(fit))

  nd <- data.frame(wt = mean(mtcars$wt), am = c(0, 1))
  ytilde <- posterior_predict(fit, newdata = nd)
  all.equal(ncol(ytilde), nrow(nd))
}
}

# Also see help("posterior_predict", package = "rstanarm")

}

