% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_doc.R
\docType{data}
\name{senate114}
\alias{senate114}
\title{Rollcall vote data for 114th Senate}
\format{A rollcall vote object that is a list with \code{votes} as the legislator-bill rollcall vote matrix.}
\source{
\url{http://www.voteview.com}
}
\usage{
senate114
}
\description{
This rollcall vote object (see \code{\link[pscl]{rollcall}}) contains the full voting records for the 114th Senate in the US
Congress. The data can be pre-processed via the \code{\link{id_make}} function for estimation. See package vignette for details.
}
\keyword{datasets}
