% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\name{stan_trace,idealstan-method}
\alias{stan_trace,idealstan-method}
\title{Plot the MCMC posterior draws by chain}
\usage{
\S4method{stan_trace}{idealstan}(object, par = "L_full[1]")
}
\arguments{
\item{object}{A fitted \code{idealstan} model}

\item{par}{The character string  name of a parameter in the model}

\item{...}{Other options passed on to \code{\link[bayesplot]{mcmc_trace}}}
}
\description{
This function allows you to produce trace plots for assessing the quality
and convergence of MCMC chains.
}
\details{
To use this function, you must pass a fitted \code{idealstan} object
along with the name of a parameter in the model. To determine these
parameter names, use the \code{summary} function or obtain the data
from a plot by passing the \code{return_data=TRUE} option to 
\code{id_plog_legis} or \code{id_plot_legis_dyn} to find the 
name of the parameter in the Stan model.

This function is a simple wrapper around \code{\link[bayesplot]{mcmc_trace}}. 
Please refer to that function's documentation for further options.
}
