% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\docType{methods}
\name{summary,idealstan-method}
\alias{summary,idealstan-method}
\title{Posterior Summaries for fitted \code{idealstan} object}
\usage{
\S4method{summary}{idealstan}(object, pars = NULL)
}
\arguments{
\item{object}{An \code{idealstan} object fitted by \code{\link{id_estimate}}}

\item{pars}{A character string of the name of the parameter in the Stan model}
}
\value{
A \code{\link[dplyr]{tibble}} data frame with parameters as rows and descriptive statistics as columns
}
\description{
This function produces quantiles and standard deviations for the posterior samples of \code{idealstan} objects.
}
